requirements.txt
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
easyocr>=1.7.0
requests>=2.31.0
httpx>=0.25.0
mediapipe>=0.10.0
PyYAML>=6.0.1
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
